# api/openapi.yaml
openapi: 3.0.3
info:
  title: TimeBeacon Intelligent Time Tracking API
  description: |
    Production-ready API for intelligent time tracking with LLM-powered duration estimation,
    project matching, and automated data import from email, calendar, and document sources.
  version: 1.0.0
  contact:
    name: TimeBeacon API Support
    email: api@timebeacon.io
  license:
    name: Proprietary
servers:
  - url: https://api.timebeacon.io/v1
    description: Production server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /time-entries:
    get:
      tags: [Time Entries]
      summary: List time entries
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TimeEntryStatus'
      responses:
        '200':
          description: Time entries retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntriesResponse'

    post:
      tags: [Time Entries]
      summary: Create a new time entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeEntryRequest'
      responses:
        '201':
          description: Time entry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'
  /time-entries/{id}:
    get:
      tags: [Time Entries]
      summary: Get a specific time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Time entry retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'

    put:
      tags: [Time Entries]
      summary: Update a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTimeEntryRequest'
      responses:
        '200':
          description: Time entry updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntry'

  # AI Processing
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TimeEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
          nullable: true
        client_id:
          type: string
          format: uuid
          nullable: true
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        duration_minutes:
          type: integer
          nullable: true
        title:
          type: string
          maxLength: 500
        description:
          type: string
        activity_type:
          $ref: '#/components/schemas/ActivityType'
        confidence_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        ai_summary:
          type: string
          nullable: true
        source_type:
          $ref: '#/components/schemas/SourceType'
        status:
          $ref: '#/components/schemas/TimeEntryStatus'
        is_billable:
          type: boolean
        is_ai_generated:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ActivityType:
      type: string
      enum: [email, meeting, document, call, manual]
    SourceType:
      type: string
      enum: [gmail, calendar, drive, zoom, manual]
    TimeEntryStatus:
      type: string
      enum: [active, completed, billed, archived]

    # Request/Response Bodies
    CreateTimeEntryRequest:
      type: object
      required: [start_time, activity_type]
      properties:
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        title:
          type: string
          maxLength: 500
        description:
          type: string
        activity_type:
          $ref: '#/components/schemas/ActivityType'
        project_id:
          type: string
          format: uuid
        is_billable:
          type: boolean
          default: true
    UpdateTimeEntryRequest:
      type: object
      properties:
        end_time:
          type: string
          format: date-time
        title:
          type: string
          maxLength: 500
        description:
          type: string
        project_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TimeEntryStatus'
        is_billable:
          type: boolean

    # AI Processing
...
    TimeEntriesResponse:
      type: object
      properties:
        time_entries:
          type: array
          items:
            $ref: '#/components/schemas/TimeEntry'
        total_count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...